+++ aes.h
@@ -35,16 +35,16 @@
 
 struct AES_ctx
 {
-  uint8_t RoundKey[AES_keyExpSize];
+  int RoundKey[AES_keyExpSize];
 #if defined(CTR) && (CTR == 1)
-  uint8_t Iv[AES_BLOCKLEN];
+  int Iv[AES_BLOCKLEN];
 #endif
 };
 
-void AES_init_ctx(struct AES_ctx* ctx, const uint8_t* key);
+void AES_init_ctx(struct AES_ctx* ctx, const int* key);
 #if defined(CTR) && (CTR == 1)
-void AES_init_ctx_iv(struct AES_ctx* ctx, const uint8_t* key, const uint8_t* iv);
-void AES_ctx_set_iv(struct AES_ctx* ctx, const uint8_t* iv);
+void AES_init_ctx_iv(struct AES_ctx* ctx, const int* key, const int* iv);
+void AES_ctx_set_iv(struct AES_ctx* ctx, const int* iv);
 #endif
 
 #if defined(CTR) && (CTR == 1)
